Added: log4j2 logger, QuestionsInitializer, AnswersInitializer, some initialization and logic. Refactored: pom.xml, quest.properties, etc...

+++ src/main/java/ru/javarush/module3/quest/repository/QuestionRepository.java
@@ -12,7 +12,7 @@ public class QuestionRepository {
         this.idToQuestion = idToQuestion;
     }
 
-    Optional<Question> findById(Long id) {
+    Optional<Question> findById(int id) {
         return Optional.ofNullable(idToQuestion.get(id));
     }
 }
